% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heckitmfx_level.R
\name{heckitmfx_level}
\alias{heckitmfx_level}
\title{Marginal Effects for Two-Step Sample Selection Models (Heckit) with Outcome Variable in Level}
\usage{
heckitmfx_level(data, selection, outcome)
}
\arguments{
\item{data}{The dataset}

\item{selection}{The selection equation}

\item{outcome}{The outcome equation}
}
\value{
A dataframe with estimated parameters, partial and total marginal effects
}
\description{
Partial and total marginal effects for two-step sample selection models (Heckit) with outcome variable in level. Partial marginal effects refer to marginal effects associated with changes in the probability of selection and with changes in the conditional outcome. Total marginal effects refer to marginal effects on the unconditional outcome.
}
\examples{
# Example following Hoffmann & Kassouf (2005) specification (eqs. 1 & 2)
\dontrun{heckitmfx_log(data, L ~ z1 + z2, log(g) ~ x1 + x2)}

# Example using simulated data of household tourism expenditures (in level, i.e. monetary units)
data(tourexp)
selection <- "participation ~ income + education + health"
outcome <- "expenditure ~ income + education + tripweather"
heckitmfx_level(tourexp, selection, outcome)

}
\references{
Hoffmann, R. & Kassouf, A. L. (2005) Deriving conditional and unconditional marginal effects in log earnings equations estimated by Heckman's procedure, Applied Economics, 37:11, 1303-1311, DOI: \href{https://www.doi.org/10.1080/00036840500118614}{https://www.doi.org/10.1080/00036840500118614}. (Equation A9).
}
