% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tourexp}
\alias{tourexp}
\title{Sample data documentation}
\format{
A dataframe with 1000 rows and 6 variables:
\describe{
\item{income}{Household income}
\item{education}{Education level of the household head}
\item{health}{Health status index of the household members}
\item{tripweather}{Weather quality in the destination of the trip}
\item{participation}{Dummy variable for tourism participation}
\item{expenditure}{Total household tourism expenditure}

data(tourexp)
}
}
\usage{
tourexp
}
\description{
Simulated survey data about household tourism demand
}
\keyword{datasets}
